HMMPlotSerie(RHmm)	R Documentation

Plot univariates series in each estimated states

Description

This function plots the time series in each hidden states.

Usage


    HMMPlotSerie(obs, states, dis="NORMAL", color="green")
    

Arguments

obs 	 the vector, list of vectors of observations

states 	 a ViterbiClass object which gives the hidden states or a vector or a lis of vectors of integer 1 to the number of hidden states

dis 	 Distribution name = 'NORMAL', 'DISCRETE', 'MIXTURE'. Default 'NORMAL'.

color 	 color for the kernel density plot

Value

none

Note

HMMPlotSerie is not implemented for multivariate observations.

The time series of observations for each hidden states of the model are plotted using:
plot(obs[states=i])) and i in 1..max(States)

See Also

viterbi, plot

Examples


data(Cac40)

data(geyser)
obs <- geyser$duration
#Fits an 3 states gaussian model for geyser duration
ResFitGeyser <- HMMFit(obs, nStates=3)
VitGeyser <- viterbi(ResFitGeyser, obs)
#plot the series
HMMPlotSerie(obs, VitGeyser)

[Package RHmm version 0.9.0 Index]